CometRewards with multiplier (#702)
* Initial implementation of CometRewards with multiplier

* Simply implementation a bit

* Refactor slightly for more readability

* Update unit tests

* Implement setRewardsClaimed (#703)

This PR introduces a `setRewardsClaimed` function that gives the admin the ability to set rewards claimed for any user. This gives us the ability to initialize a new reward distribution that doesn't necessarily have to retroactively reward all past usage of Comet. 

We would have to compute off-chain the amount that each user would be able to claim at a specific block number `x` and write it on-chain via the `setRewardsClaimed` function. This allows us to essentially set the start time of the rewards accrual to be after that specific block number `x`.

* Add scenarios for testing reward multipliers

* Re-implement multiplier logic by storing multiplier in RewardConfig

* Add a unit test to sanity check zeroing out retroactive rewards for users